requirements.txt
pandas>=1.5
psycopg2-binary>=2.9
pyyaml>=6.0
pyarrow>=10.0
streamlit>=1.25
